+++ packages/types/src/catalog.ts
@@ -4,7 +4,7 @@ export type Catalog =
   Flowglad.CustomerRetrieveBillingResponse['catalog']
 
 export type Product =
-  Flowglad.CustomerRetrieveBillingResponse.Catalog.Product.Product
+  Flowglad.CustomerRetrieveBillingResponse.Catalog.Product
 
 export type SinglePaymentPrice =
   Flowglad.CustomerRetrieveBillingResponse.Catalog.Product.SinglePaymentPrice
@@ -12,4 +12,10 @@ export type SinglePaymentPrice =
 export type SubscriptionPrice =
   Flowglad.CustomerRetrieveBillingResponse.Catalog.Product.SubscriptionPrice
 
-export type Price = SinglePaymentPrice | SubscriptionPrice
+export type UsagePrice =
+  Flowglad.CustomerRetrieveBillingResponse.Catalog.Product.UsagePrice
+
+export type Price =
+  | SinglePaymentPrice
+  | SubscriptionPrice
+  | UsagePrice