+++ packages/express/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @flowglad/express
 
+## 0.8.13
+
+### Patch Changes
+
+- Add subscription.current, checkoutSession.quantity
+- Updated dependencies
+  - @flowglad/server@0.8.13
+  - @flowglad/shared@0.8.13
+  - @flowglad/react@0.8.13
+
 ## 0.8.12
 
 ### Patch Changes