+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @flowglad/shared
 
+## 0.8.13
+
+### Patch Changes
+
+- Add subscription.current, checkoutSession.quantity
+
 ## 0.8.12
 
 ### Patch Changes